fix: Cast DB_PORT to integer for mysqli constructor
- mysqli::__construct() requires port as int, not string
- DB_PORT from config was string '3306', causing type error
- Add (int) cast to ensure correct type

+++ core/Database.php
@@ -12,7 +12,7 @@ class Database {
 
     private function __construct() {
         $host = DB_HOST;
-        $port = DB_PORT;
+        $port = (int) DB_PORT;
         
         if ($host === 'localhost') {
             $host = '127.0.0.1';
